requirements.txt
# Core
Django==4.2.4

# Model Tools
django-simple-history==3.3.0
django-model-utils==4.3.1

# Rest apis
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework-simplejwt== 5.3.0

# Rest apis Permissions
drf-access-policy==1.5.0

# Developer Tools
ipdb==0.13.13
ipython==8.14.0
django-extensions==3.2.3

# other required packages
django-environ==0.8.1

# AWS
boto3==1.28.35

# static file
django-compressor==4.4